HP-UX and Tru64 fixes
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@305 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_pcstep.c
@@ -1105,7 +1105,7 @@ mxd_pcstep_constant_velocity_move( MX_MOTOR *motor )
 	if ( motor->constant_velocity_move >= 0 ) {
 		direction = 1;
 	} else {
-		direction = -1;
+		direction = (uint32_t) -1;
 	}
 
 	mx_status = mxi_pcstep_command( pcstep,
